requirements.txt
asgiref==3.6.0
Django==4.1.6
gunicorn==20.1.0
polars==0.16.2
psycopg2-binary==2.9.5
sqlparse==0.4.3
asgiref==3.6.0
Django==4.1.6
gunicorn==20.1.0
polars==0.16.2
psycopg2-binary==2.9.5
sqlparse==0.4.3
whitenoise==6.3.0
django-cors-headers==3.13.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-cors-headers==3.13.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
openai==0.26.5
polars==0.16.2
psycopg2-binary==2.9.5
requests==2.28.2
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.14
whitenoise==6.3.0
yarl==1.8.2
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-cors-headers==3.13.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
multidict==6.0.4
openai==0.26.5
polars==0.16.2
psycopg2-binary==2.9.5
python-dotenv==0.21.1
requests==2.28.2
sqlparse==0.4.3
tqdm==4.64.1
urllib3==1.26.14
whitenoise==6.3.0
yarl==1.8.2
